+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQY7B5AEA62HYPFAAQQTV49D"
+        "01KQYDJQJJT8E65WRDJRFEQ0FQ"
     )
     r"""The unique identifier of the data source"""