+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-06T08:44:02.209Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-06T10:33:02.161Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")